NavBar fora do APP e como componente
